+++ src/main/java/blackjack/Application.java
@@ -1,5 +1,7 @@
 package blackjack;
 
+import blackjack.controller.BlackJackController;
+
 public class Application {
     public static void main(String[] args) {
         BlackJackController blackJackController = new BlackJackController();
