+++ src/main/java/blackjack/controller/BlackJackController.java
@@ -1,5 +1,13 @@
-package blackjack;
+package blackjack.controller;
 
+import blackjack.domain.BlackJackGame;
+import blackjack.domain.Dealer;
+import blackjack.domain.Deck;
+import blackjack.domain.GameResult;
+import blackjack.view.InputView;
+import blackjack.view.OutputView;
+import blackjack.domain.Player;
+import blackjack.domain.Players;
 import java.util.HashMap;
 import java.util.List;
