+++ src/main/java/blackjack/domain/GameResult.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 public enum GameResult {
     WIN("승"),
