+++ src/main/java/blackjack/domain/Player.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 public class Player {
