+++ src/test/java/blackjack/domain/CardPatternTest.java
@@ -1,9 +1,8 @@
-package blackjack;
+package blackjack.domain;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
