+++ src/test/java/blackjack/domain/GameResultTest.java
@@ -1,8 +1,7 @@
-package blackjack;
+package blackjack.domain;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
